Fix TypeLoadException on non-Windows platoforms
If you attempt to use the TaskbarProgress class on non-Windows platforms, it will throw a TypeLoadException stating that "COM is not supported". This is presumably because of the use of the [ComImport] attribute in TaskbarProgress.

This modifies the EnableTaskBarProgress property to throw an exception if we try to set it to true on non-Windows platforms.

It also fixes the demo to conditionally set this property based on the runtime platform.