Performance improvements (#28)
* Performance improvements

Apparently, the use of Console.CursorTop and Console.CursorLeft is *very* expensive on .NET Core 2.1 on Linux. As in "I left the demo program running for 12 hours and it was still on 0% overall progress" kind of expensive.

This changes ProgressBar to keep track of the cursor position internally instead of using the Console.CursorTop getter. Also, calls to the setters of Console.CursorTop and Console.CursorLeft are replaced with Console.SetCursorPosition() which has much better performance.

The demo program can now run in seconds instead of days.

* uncomment all examples in example project